lnwallet: rewrite validateCommitmentSanity to account for constraints.
This commit rewrites the way validateCommitmentSanity checks
that all channel constraints are validated. evaluateHTLCView
is made optionally non-mutable, such that it can be used to
calculate the resulting balances from applying the given
updates. Then the channel constraints are validated from
the point-of-view of the local and the remote node, to make
sure both will accept the updates.